requirements.txt
--find-links https://tortall.net/~robotpy/wheels/2023/raspbian/
--prefer-binary

asyncio
autobahn_client
colorama
dataclasses
filterpy
fix-protobuf-imports
flask
netifaces
numpy
peewee
protobuf
psutil
pyapriltags
pydantic
pytest
robotpy-cscore
scipy
zeroconf
thrift
opencv-python
requests
ultralytics